refactor/#8: Modify jwt directory path
