+++ src/apis/auth/commands/generate-access-token/generate-access-token.command-handler.ts
@@ -1,8 +1,8 @@
 import { Inject } from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { GenerateAccessTokenCommand } from '@src/apis/auth/commands/generate-access-token/generate-access-token.command';
-import { AppJwtServicePort } from '@src/libs/jwt/services/app-jwt.service-port';
-import { APP_JWT_SERVICE_DI_TOKEN } from '@src/libs/jwt/tokens/app-jwt.di-token';
+import { AppJwtServicePort } from '@src/libs/app-jwt/services/app-jwt.service-port';
+import { APP_JWT_SERVICE_DI_TOKEN } from '@src/libs/app-jwt/tokens/app-jwt.di-token';
 
 @CommandHandler(GenerateAccessTokenCommand)
 export class GenerateAccessTokenCommandHandler
