+++ src/libs/app-jwt/services/app-jwt.service-port.ts
@@ -1,4 +1,4 @@
-import type { JwtPayload } from '@src/libs/jwt/types/app-jwt.interface';
+import type { JwtPayload } from '@src/libs/app-jwt/types/app-jwt.interface';
 
 export interface AppJwtServicePort {
   generateAccessToken(payload: JwtPayload): Promise<string>;
