+++ src/libs/app-jwt/services/app-jwt.service.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { JwtService } from '@nestjs/jwt';
 import { HttpUnauthorizedException } from '@src/libs/exceptions/client-errors/exceptions/http-unauthorized.exception';
 import { COMMON_ERROR_CODE } from '@src/libs/exceptions/types/errors/common/common-error-code.constant';
-import { AppJwtServicePort } from '@src/libs/jwt/services/app-jwt.service-port';
-import type { JwtPayload } from '@src/libs/jwt/types/app-jwt.interface';
+import { AppJwtServicePort } from '@src/libs/app-jwt/services/app-jwt.service-port';
+import type { JwtPayload } from '@src/libs/app-jwt/types/app-jwt.interface';
 
 @Injectable()
 export class AppJwtService implements AppJwtServicePort {
